// src/app/countries/interfaces/country.interface.ts
// Generated by https://quicktype.io

export enum Region {
  Africa = 'Africa',
  Americas = 'Americas',
  Asia = 'Asia',
  Europe = 'Europe',
  Oceania = 'Oceania'
}

export interface ISmallCountry {
  name: string,
  cca3: string,
  borders: string[];
}

export interface ICountry {
  name:         IName;
  cca3:         string;
  status:       Status;
  unMember:     boolean;
  currencies:   { [key: string]: ICurrency };
  idd:          IIdd;
  capital:      string[];
  altSpellings: string[];
  region:       string[];
  subregion:    Subregion;
  languages:    { [key: string]: string };
  translations: { [key: string]: ITranslation };
  latlng:       number[];
  landlocked:   boolean;
  area:         number;
  demonyms:     IDemonyms;
  flag:         string;
  maps:         IMaps;
  population:   number;
  fifa?:        string;
  car:          ICar;
  timezones:    string[];
  continents:   Continent[];
  flags:        IFlags;
  coatOfArms:   ICoatOfArms;
  startOfWeek:  StartOfWeek;
  capitalInfo:  ICapitalInfo;
  postalCode?:  IPostalCode;
  borders?:     string[];
  gini?:        { [key: string]: number };
}

export interface ICapitalInfo {
  latlng?: number[];
}

export interface ICar {
  signs?: string[];
  side:   ISide;
}

export enum ISide {
  Left = "left",
  Right = "right",
}

export interface ICoatOfArms {
  png?: string;
  svg?: string;
}

export enum Continent {
  NorthAmerica = "North America",
  Oceania = "Oceania",
  SouthAmerica = "South America",
}

export interface ICurrency {
  name:   string;
  symbol: string;
}

export interface IDemonyms {
  eng:  IEng;
  fra?: IEng;
}

export interface IEng {
  f: string;
  m: string;
}

export interface IFlags {
  png:  string;
  svg:  string;
  alt?: string;
}

export interface IIdd {
  root:     string;
  suffixes: string[];
}

export interface IMaps {
  googleMaps:     string;
  openStreetMaps: string;
}

export interface IName {
  common:     string;
  official:   string;
  nativeName: { [key: string]: ITranslation };
}

export interface ITranslation {
  official: string;
  common:   string;
}

export interface IPostalCode {
  format: string;
  regex?: string;
}



export enum StartOfWeek {
  Monday = "monday",
  Sunday = "sunday",
}

export enum Status {
  OfficiallyAssigned = "officially-assigned",
}

export enum Subregion {
  Caribbean = "Caribbean",
  CentralAmerica = "Central America",
  NorthAmerica = "North America",
  SouthAmerica = "South America",
}